cmd/observe: Refactor into own package
