+++ cmd/observe/observe_filter_test.go
@@ -12,20 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cmd
+package observe
 
 import (
 	"testing"
 
+	pb "github.com/cilium/hubble/api/v1/flow"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	pb "github.com/cilium/hubble/api/v1/flow"
+	"go.uber.org/zap"
 )
+
+var nopl = zap.NewNop()
 
 func TestNoBlacklist(t *testing.T) {
 	f := newObserveFilter()
-	cmd := newObserverCmd(f)
+	cmd := newObserveCmd(nopl, f)
 
 	require.NoError(t, cmd.Flags().Parse([]string{
 		"--from-ip", "1.2.3.4",
@@ -35,7 +37,7 @@ func TestNoBlacklist(t *testing.T) {
 
 func TestDefaultTypes(t *testing.T) {
 	f := newObserveFilter()
-	cmd := newObserverCmd(f)
+	cmd := newObserveCmd(nopl, f)
 
 	require.NoError(t, cmd.Flags().Parse([]string{}))
 	assert.Equal(t, []*pb.FlowFilter{{
@@ -48,7 +50,7 @@ func TestDefaultTypes(t *testing.T) {
 
 func TestConflicts(t *testing.T) {
 	f := newObserveFilter()
-	cmd := newObserverCmd(f)
+	cmd := newObserveCmd(nopl, f)
 
 	err := cmd.Flags().Parse([]string{
 		"--from-ip", "1.2.3.4",
@@ -64,7 +66,7 @@ func TestConflicts(t *testing.T) {
 
 func TestTrailingNot(t *testing.T) {
 	f := newObserveFilter()
-	cmd := newObserverCmd(f)
+	cmd := newObserveCmd(nopl, f)
 
 	err := cmd.Flags().Parse([]string{
 		"--from-ip", "1.2.3.4",
@@ -72,14 +74,14 @@ func TestTrailingNot(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	err = handleArgs(cmd, f)
+	err = handleArgs(nopl, cmd, f)
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "trailing --not")
 }
 
 func TestFilterDispatch(t *testing.T) {
 	f := newObserveFilter()
-	cmd := newObserverCmd(f)
+	cmd := newObserveCmd(nopl, f)
 
 	require.NoError(t, cmd.Flags().Parse([]string{
 		"--from-ip", "1.2.3.4",
@@ -90,7 +92,7 @@ func TestFilterDispatch(t *testing.T) {
 		"-t", "l7", // int:129 in cilium-land
 	}))
 
-	require.NoError(t, handleArgs(cmd, f))
+	require.NoError(t, handleArgs(nopl, cmd, f))
 
 	assert.Equal(t, []*pb.FlowFilter{
 		{
@@ -109,7 +111,7 @@ func TestFilterDispatch(t *testing.T) {
 
 func TestFilterLeftRight(t *testing.T) {
 	f := newObserveFilter()
-	cmd := newObserverCmd(f)
+	cmd := newObserveCmd(nopl, f)
 
 	require.NoError(t, cmd.Flags().Parse([]string{
 		"--ip", "1.2.3.4",
@@ -121,7 +123,7 @@ func TestFilterLeftRight(t *testing.T) {
 		"--http-status", "200",
 	}))
 
-	require.NoError(t, handleArgs(cmd, f))
+	require.NoError(t, handleArgs(nopl, cmd, f))
 
 	assert.Equal(t, []*pb.FlowFilter{
 		{
@@ -150,7 +152,7 @@ func TestFilterLeftRight(t *testing.T) {
 
 func TestLabels(t *testing.T) {
 	f := newObserveFilter()
-	cmd := newObserverCmd(f)
+	cmd := newObserveCmd(nopl, f)
 
 	err := cmd.Flags().Parse([]string{
 		"--label", "k1=v1,k2=v2",
